feat: rename old SearchBar.tsx to RoomSearchBar.tsx and use custom SearchBar
