--- fiuba-guide/src/components/SearchBar.css
@@ -1,16 +1,0 @@
-.floor-mini svg {
-  max-width: 50px;
-  max-height: 50px;
-}
-
-.result-cointainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.search-bar {
-  max-width: 400px;
-  margin: 1rem;
-  width: 100%;
-}
